Migrate the underlying test framework from `kuttl` to `chainsaw`

+++ internal/runner.go
@@ -5,7 +5,7 @@
 package internal
 
 import (
-	"fmt"
+	"log"
 	"os"
 
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
@@ -15,11 +15,13 @@ import (
 
 // RunTest runs the specified automated test
 func RunTest(o *config.AutomatedTest) error {
-	defer func() {
-		if err := os.RemoveAll(o.Directory); err != nil {
-			fmt.Println(fmt.Sprint(err, "cannot clean the test directory"))
-		}
-	}()
+	if !o.RenderOnly {
+		defer func() {
+			if err := os.RemoveAll(o.Directory); err != nil {
+				log.Printf("Cannot clean the test directory: %s\n", err.Error())
+			}
+		}()
+	}
 
 	// Read examples and inject data source values to manifests
 	manifests, err := newPreparer(o.ManifestPaths, withDataSource(o.DataSourcePath), withTestDirectory(o.Directory)).prepareManifests()
